# Backend/Syslab/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "pablokauan.tech@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/projetos": {
            "get": {
                "tags": [
                    "Projeto"
                ],
                "summary": "Listar projetos",
                "operationId": "e6d07ef7e5421002718e665b7a2e9fc1",
                "responses": {
                    "200": {
                        "description": "Lista de projetos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Projeto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projeto"
                ],
                "summary": "Criar projeto",
                "operationId": "440ca35050f45e2d2504fc1fb3ed3d55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "date_begin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_end": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projeto criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projeto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projetos/{id}": {
            "get": {
                "tags": [
                    "Projeto"
                ],
                "summary": "Buscar projeto por ID",
                "operationId": "acfba0f375881d8905f24fbcd6a04c6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projeto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projeto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projeto"
                ],
                "summary": "Atualizar projeto",
                "operationId": "09078063a88fa0ac37e99120fa6af3db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "date_begin": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_end": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projeto atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projeto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projeto"
                ],
                "summary": "Remover projeto",
                "operationId": "9caf5dc4c5a5704f6ab83466dad9d396",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Projeto removido"
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "News": {
                "title": "News",
                "description": "Notícia publicada no sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Novo laboratório é inaugurado"
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Espaço amplia pesquisas na universidade",
                        "nullable": true
                    },
                    "body_text": {
                        "type": "string",
                        "example": "O laboratório foi inaugurado com foco em inovação..."
                    },
                    "image": {
                        "type": "string",
                        "example": "news.jpg",
                        "nullable": true
                    },
                    "datastamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-01T10:30:00Z"
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "$ref": "#/components/schemas/Pessoa"
                    }
                },
                "type": "object"
            },
            "Pessoa": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pablo Kauan"
                    },
                    "image": {
                        "type": "string",
                        "example": "perfil.jpg",
                        "nullable": true
                    },
                    "research": {
                        "type": "string",
                        "example": "Inteligência Artificial",
                        "nullable": true
                    },
                    "lattes": {
                        "type": "string",
                        "nullable": true
                    },
                    "orcid": {
                        "type": "string",
                        "nullable": true
                    },
                    "descricao": {
                        "type": "string",
                        "nullable": true
                    },
                    "cargo": {
                        "type": "string",
                        "nullable": true
                    },
                    "curso": {
                        "type": "string",
                        "nullable": true
                    },
                    "projetos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Projeto"
                        }
                    },
                    "news": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/News"
                        }
                    }
                },
                "type": "object"
            },
            "Projeto": {
                "title": "Projeto",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sistema de Monitoramento Energético"
                    },
                    "image": {
                        "type": "string",
                        "example": "projeto.jpg",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Projeto voltado ao monitoramento de consumo energético",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "ativo"
                    },
                    "date_begin": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31",
                        "nullable": true
                    },
                    "pivot": {
                        "properties": {
                            "integrantes": {
                                "type": "string",
                                "example": "Desenvolvedor Backend",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "pessoas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Pessoa"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}